Added version symlinks: 0.x -> v0.x for compat
Previously, URLs on docs.janusgraph.org were of the form:

* https://docs.janusgraph.org/0.1/
* https://docs.janusgraph.org/0.2/
* etc.

and with the new versions having the `v` prefix, all of those URLs are now
broken. This fixes it for all past docs (including docs that still need to be
published, such as the `0.1` and `0.4` series), as well as the upcoming version
`0.5`.

Signed-off-by: Misha Brukman <mbrukman@google.com>